requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.6
cycler==0.11.0
Django==3.2.5
django-bootstrap-form==3.4
django-crispy-forms==1.12.0
django-phonenumber-field==6.0.0
django-registration==3.2
django-survey-and-report==1.3.35
django-tastypie==0.14.4
djangorestframework==3.13.0
fonttools==4.28.3
greenlet==1.1.2
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.4
ordereddict==1.1
packaging==21.3
pandas==1.3.5
phonenumbers==8.12.39
Pillow==8.3.1
pyparsing==3.0.6
pySankeyBeta==1.3.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2021.1
PyYAML==6.0
requests==2.26.0
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.28
sqlparse==0.4.1
urllib3==1.26.7
